--- dataconnect/buildSrc/src/main/kotlin/com/google/firebase/example/dataconnect/gradle/LocalConfig.kt
@@ -1,31 +1,0 @@
-/*
- * Copyright 2024 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.firebase.example.dataconnect.gradle
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-internal data class LocalConfig(
-    val npmExecutable: String? = null,
-    val nodeExecutable: String? = null
-) :
-    java.io.Serializable {
-    companion object {
-        @Suppress("ConstPropertyName")
-        private const val serialVersionUID = 6103369929496556758L
-    }
-}
