+++ dataconnect/buildSrc/src/main/kotlin/com/google/firebase/example/dataconnect/gradle/tasks/util.kt
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package com.google.firebase.example.dataconnect.gradle
+package com.google.firebase.example.dataconnect.gradle.tasks
 
-import java.io.File
 import org.gradle.api.Task
 import org.gradle.process.ExecSpec
+import java.io.File
 
 internal fun Task.runCommand(logFile: File, configure: ExecSpec.() -> Unit) {
     val effectiveLogFile = if (logger.isInfoEnabled) null else logFile
